9:15 14/5/2023

+++ src/main/java/poly/dao/ProductDao.java
@@ -88,7 +88,13 @@ public class ProductDao {
 		Session session = factory.openSession();
 		Transaction t = session.beginTransaction();
 		Message message = new Message();
+		if (p.getInventoryCapability().toArray().length > 0) {
+			message.setType("error");
+			message.setContent("Sản phẩm đã có trong kho hàng. Xóa thất bại!");
+			return message;
+		}
 		try {
+			p.setInventoryCapability(null);
 			session.delete(p);
 			t.commit();
 			message.setType("success");
